CrashBucketing/ClientApp/src/app/nav-menu/nav-menu.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { NavMenuComponent } from './nav-menu.component';
import { RouterTestingModule } from '@angular/router/testing';

describe('NavMenuComponent', () => {
  let component: NavMenuComponent;
  let fixture: ComponentFixture<NavMenuComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ NavMenuComponent ],
      imports: [ RouterTestingModule ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(NavMenuComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should have a title of CrashBucketing', async(() => {
    const app = fixture.debugElement.componentInstance;
    expect(app.title).toEqual('CrashBucketing');
  }));

  it('Nav bar should not be expanded on load', async(() => {
    const app = fixture.debugElement.componentInstance;
    expect(app.isExpanded).toEqual(false);
  }));
});

